+++ pkg/grpc/default_api/api_state.go
@@ -18,14 +18,10 @@ import (
 	"github.com/dapr/components-contrib/state"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/protobuf/types/known/emptypb"
-	"mosn.io/layotto/pkg/common"
 	dapr_common_v1pb "mosn.io/layotto/pkg/grpc/dapr/proto/common/v1"
 	dapr_v1pb "mosn.io/layotto/pkg/grpc/dapr/proto/runtime/v1"
-	state2 "mosn.io/layotto/pkg/runtime/state"
 	runtimev1pb "mosn.io/layotto/spec/proto/runtime/v1"
 	_ "net/http/pprof"
-
-	"google.golang.org/grpc/status"
 )
 
 func (a *api) SaveState(ctx context.Context, in *runtimev1pb.SaveStateRequest) (*emptypb.Empty, error) {
